requirements.txt
requests==2.27.1
dhooks==1.1.4
ujson==5.2.0
git+https://github.com/Rapptz/discord.py
python-dotenv==0.20.0
psutil==5.9.0
praw==7.6.0
APScheduler==3.9.1
ujson==5.2.0
dhooks==1.1.4
asyncpraw==7.5.0